# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ClassicPOS API",
        "description": "ClassicPOS — Point of Sale REST API. Multi-tenant SaaS with role-based access control.",
        "contact": {
            "name": "ClassicPOS Support",
            "email": "support@oakitsolutionsandsupplies.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/api/v1/plans": {
            "get": {
                "tags": [
                    "Public Plans"
                ],
                "summary": "List active subscription plans with pricing",
                "operationId": "97359f51d0fe8c07942f0bbcd0c38421",
                "parameters": [
                    {
                        "name": "currency",
                        "in": "query",
                        "description": "Currency code for pricing",
                        "schema": {
                            "type": "string",
                            "default": "USD"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plans with localized pricing"
                    }
                }
            }
        },
        "/api/v1/plans/{slug}": {
            "get": {
                "tags": [
                    "Public Plans"
                ],
                "summary": "Get a single plan by slug",
                "operationId": "850d4f07bac427bd5e52e4845fa714e1",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "currency",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "USD"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plan details"
                    },
                    "404": {
                        "description": "Plan not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Sanctum token from login endpoint",
                "scheme": "bearer"
            }
        }
    },
...
}